refactor(organization): remove alias, types and rename files
