+++ packages/pcgl-validation/src/index.ts
@@ -17,5 +17,6 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export * from './types.js';
-export * from './validationSchemas.js';
+export * from './common/strings.js';
+export * from './schemas.js';
+export * from './utils/regex.js';
