+++ packages/pcgl-validation/tsconfig.json
@@ -1,9 +1,4 @@
 {
 	"extends": "../../tsconfig.base.json",
-	"include": ["src"],
-	"compilerOptions": {
-		"paths": {
-			"@/*": ["./src/*"]
-		}
-	}
+	"include": ["src"]
 }
